refactorize ExplorationTaskGroup. provide abs method make_task. provide ExplorationTaskGroupData as a data holder
