+++ dpgen2/exploration/task/__init__.py
@@ -23,5 +23,8 @@ from .stage import (
 )
 from .task import (
     ExplorationTask,
+)
+from .task_group import (
     ExplorationTaskGroup,
+    ExplorationTaskGroupData,
 )
