+++ tests/test_prep_run_lmp.py
@@ -70,7 +70,7 @@ from dpgen2.constants import (
 )
 from dpgen2.exploration.task import (
     ExplorationTask,
-    ExplorationTaskGroup,
+    ExplorationTaskGroupData,
 )
 from dpgen2.op.prep_lmp import (
     PrepLmp,
@@ -90,7 +90,7 @@ default_config = normalize_step_dict(
 
 
 def make_task_group_list(ngrp, ntask_per_grp):
-    tgrp = ExplorationTaskGroup()
+    tgrp = ExplorationTaskGroupData()
     for ii in range(ngrp):
         for jj in range(ntask_per_grp):
             tt = ExplorationTask()
